requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-kasa>=0.7.0
tinytuya>=1.15.0
httpx>=0.27.0
pyyaml>=6.0.0
paho-mqtt>=2.1.0
aiofiles>=24.0.0
python-dotenv>=1.0.0
